# Initial implementation of a faster repr