--- a/changelog.d/819.changes.rst
+++ b/changelog.d/819.changes.rst
@@ -1,1 +1,1 @@
-The generated `__repr__` is significantly faster on Pythons with F-strings.
+The generated ``__repr__`` is significantly faster on Pythons with F-strings.