--- a/src/attr/_compat.py
+++ b/src/attr/_compat.py
@@ -8,7 +8,11 @@ import warnings
 
 PY2 = sys.version_info[0] == 2
 PYPY = platform.python_implementation() == "PyPy"
-HAS_F_STRINGS = sys.version_info[:2] >= (3, 7)
+HAS_F_STRINGS = (
+    sys.version_info[:2] >= (3, 7)
+    if not PYPY
+    else sys.version_info[:2] >= (3, 6)
+)
 PY310 = sys.version_info[:2] >= (3, 10)