--- a/src/attr/_make.py
+++ b/src/attr/_make.py
@@ -2705,7 +2705,7 @@ class Attribute(object):
             type=type,
             cmp=None,
             inherited=False,
-            **inst_dict,
+            **inst_dict
         )
 
     @property