Initial implementation, CHANGELOG to follow.

Benchmark used:
```
pyperf timeit -g -s "from attr import make_class, attrib; A = make_class('A', ['a', 'b', 'c', 'd', 'e']); a = A(1, 2, 3, 4, 5)" "repr(a)"
```

Old code:
```
Mean +- std dev: 2.09 us +- 0.07 us
```

New code:
```
Mean +- std dev: 957 ns +- 27 ns
```