@hynek 2021-05-20
Looks like you’ll need a conditionally imported `repr_fstrings.py` or something.