@Tinche 2021-05-20
> Looks like you’ll need a conditionally imported `repr_fstrings.py` or something.

I think it just means I can't use f-strings while preparing the function. I can just use string concatenation or `.format()` instead, not a big deal.