@hynek 2021-05-21
y u hate Python 2 Tin