@hynek 2021-05-21
Still getting a depressing amount of Python 2 downloads 😢