@sscherfke 2021-05-21
The current versions that still support Python 2 won’t go away. :)

Many of the new features are related to Python-3-only features (e.g., typing, pattern matching) and are not benefical for Python 2 users anyways. So I don’t realy see the problem with new versions only supporting python 3.6 (or 3.7) upwards. :)