@Tinche 2021-05-21
Ok, looks like just some line length warnings which I can fix. Ready for a full review.