@hynek 2021-05-23
Looks like one more line: https://codecov.io/gh/python-attrs/attrs/src/f37fc134b66c14b03a09fc09ca59f188537e1911/src/attr/_make.py#L1976