@hynek 2021-05-24
Guess we have to add 3.6 to coverage then