@Tinche 2021-05-26
> I haven't spent super much time understanding it, because the tests are passing so it should be fine. #yolo
> 
> But I'd like to avoid snail-formatting in new code (maybe should be adding to the contribution guide), so let's not do it. :)

What's snail-formatting btw?