@hynek 2021-05-26
No, snail formatting is what I call `str.format()` because it's slow as a snail. :D 🐌

Modulo + tuple is the second fastest approach.